@@ -6,27 +6,27 @@
 
   <groupId>gov.nist.secauto.oscal.tools</groupId>
   <artifactId>server</artifactId>
-  <version>1.0.0-SNAPSHOT</version>
+  <version>1.0.0</version>
 
   <properties>
-    
+    <maven.resolver.transport.snapshot.update>never</maven.resolver.transport.snapshot.update>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
-    <kotlin.version>2.0.20</kotlin.version>
+    <kotlin.version>1.8.21</kotlin.version>
     <kotlinx.coroutines.version>1.9.0</kotlinx.coroutines.version>
 
     <maven-shade-plugin.version>3.6.0</maven-shade-plugin.version>
-    <maven-surefire-plugin.version>2.22.2</maven-surefire-plugin.version>
+    <maven-surefire-plugin.version>3.5.2</maven-surefire-plugin.version>
     <exec-maven-plugin.version>3.4.1</exec-maven-plugin.version>
     <maven-assembly-plugin.version>3.3.0</maven-assembly-plugin.version>
     <swid-maven-plugin.version>0.7.0</swid-maven-plugin.version>
   
-    <vertx.version>4.5.10</vertx.version>
+    <vertx.version>4.5.13</vertx.version>
     <swid-maven-plugin.version>0.7.0</swid-maven-plugin.version>
     <oscal-cli.version>2.4.0</oscal-cli.version>
     <metaschema-framework.version>2.1.0</metaschema-framework.version>
     <liboscal-java.version>5.1.0</liboscal-java.version>
-    <java.version>1.8</java.version>
+    <java.version>17</java.version>
     <maven.compiler.source>${java.version}</maven.compiler.source>
     <maven.compiler.target>${java.version}</maven.compiler.target>
     <dependency.log4j2.version>2.24.1</dependency.log4j2.version>
