@@ -16,6 +16,7 @@ import kotlin.io.path.appendText
 import kotlinx.coroutines.Job
 import kotlin.coroutines.CoroutineContext
 import java.nio.file.Path
+import java.nio.file.Paths
 import java.nio.file.StandardCopyOption
 import java.nio.file.Path.of
 
@@ -41,7 +42,7 @@ class RequestHandler(
             .end(response.encode())
     }
 
-    suspend fun handleQueryRequest(ctx: RoutingContext) {
+    fun handleQueryRequest(ctx: RoutingContext) {
         try {
             logger.info("Handling Query request")
             val encodedContent = ctx.queryParam("document").firstOrNull()
@@ -54,16 +55,7 @@ class RequestHandler(
             }
             
             val content = urlProcessor.processUrl(encodedContent)
-            val args = mutableListOf("metaschema", "metapath", "eval")
-            args.add("-i")
-            args.add(content)
-            args.add("-e")
-
-            args.add(expression)
-            args.add("-m")
-            args.add(module)
-            
-            val result = commandExecutor.executeCommand(args)
+            val result = commandExecutor.evaluateMetapath(Paths.get(content), expression, module)
             logger.info(result.second)
             ctx.response().putHeader("Content-Type", "application/json")
             responseHandler.sendSuccessResponse(ctx, result.first, result.second)
@@ -73,13 +65,13 @@ class RequestHandler(
         }
     }
 
-    suspend fun handleValidateRequest(ctx: RoutingContext) {
+    fun handleValidateRequest(ctx: RoutingContext) {
         try {
             logger.info("Handling Validate request")
             val encodedContent = ctx.queryParam("document").firstOrNull()
-            val encodedModule = ctx.queryParam("module").firstOrNull()
-            val constraint = ctx.queryParam("constraint")
-            val flags = ctx.queryParam("flags")
+            val flags = ctx.queryParam("flags").toSet()
+            val constraints = ctx.queryParam("constraint").map { urlProcessor.processUrl(it) }
+            val module = ctx.queryParam("module").firstOrNull()
             
             if (encodedContent == null) {
                 responseHandler.sendErrorResponse(ctx, 400, "content parameter is missing")
@@ -87,32 +79,15 @@ class RequestHandler(
             }
             
             val content = urlProcessor.processUrl(encodedContent)
+            val inputPath = Paths.get(content)
             
-            // Generate SARIF output file path
-            
-            val args = mutableListOf("validate")
-            encodedModule?.let { module ->
-                if (module == "http://csrc.nist.gov/ns/oscal/metaschema/1.0") {
-                    args[0] = "metaschema"
-                    args.add("validate")
-                } else {
-                    args[0] = "metaschema"
-                    args.add("validate-content")
-                }
-            }
-            args.add(content)
-            
-            constraint.forEach { constraint_document ->
-                args.add("-c")
-                args.add(urlProcessor.processUrl(constraint_document))
-            }
-            flags.forEach { flag ->
-                args.add(responseHandler.flagToParam(flag))
-            }
-            
-
             try {
-                val result = commandExecutor.executeCommand(args)
+                val result = commandExecutor.validateDocument(
+                    inputPath = inputPath,
+                    flags = flags,
+                    constraints = constraints.map { Paths.get(it) },
+                    module = module
+                )
                 
                 // Read SARIF output if it exists
                 val sarifContent = if (Files.exists(of(result.second))) {
@@ -134,7 +109,7 @@ class RequestHandler(
         }
     }
 
-    suspend fun handleResolveRequest(ctx: RoutingContext) {
+    fun handleResolveRequest(ctx: RoutingContext) {
         try {
             val encodedContent = ctx.queryParam("document").firstOrNull()
             if (encodedContent != null) {
@@ -146,9 +121,7 @@ class RequestHandler(
                 // Generate output file path
                 val outputFile = of(oscalDir.toString(), "resolve-${System.nanoTime()}.tmp")
                 
-                // Build command: resolve-profile --to=FORMAT source-file destination-file
-                val args = mutableListOf("resolve-profile", "--to=$format", content, outputFile.toString())
-                val result = commandExecutor.executeCommand(args)
+                val result = commandExecutor.resolveProfile(Paths.get(content), outputFile, format)
                 
                 // Read output and clean up
                 val output = if (Files.exists(outputFile)) Files.readString(outputFile) else result.second
@@ -161,12 +134,12 @@ class RequestHandler(
                 responseHandler.sendErrorResponse(ctx, 400, "content parameter is missing")
             }
         } catch (e: Exception) {
-            logger.error("Error handling CLI request", e)
+            logger.error("Error handling request", e)
             responseHandler.sendErrorResponse(ctx, 500, "Internal server error")
         }
     }
 
-    suspend fun handleConvertRequest(ctx: RoutingContext) {
+    fun handleConvertRequest(ctx: RoutingContext) {
         try {
             val encodedContent = ctx.queryParam("document").firstOrNull()
             if (encodedContent != null) {
@@ -178,9 +151,7 @@ class RequestHandler(
                 // Generate output file path
                 val outputFile = of(oscalDir.toString(), "convert-${System.nanoTime()}.tmp")
                 
-                // Build command: convert --to=FORMAT source-file destination-file
-                val args = mutableListOf("convert", "--to=$format", content, outputFile.toString())
-                val result = commandExecutor.executeCommand(args)
+                val result = commandExecutor.convertDocument(Paths.get(content), outputFile, format)
                 
                 // Read output and clean up
                 val output = if (Files.exists(outputFile)) Files.readString(outputFile) else result.second
@@ -193,7 +164,7 @@ class RequestHandler(
                 responseHandler.sendErrorResponse(ctx, 400, "content parameter is missing")
             }
         } catch (e: Exception) {
-            logger.error("Error handling CLI request", e)
+            logger.error("Error handling request", e)
             responseHandler.sendErrorResponse(ctx, 500, "Internal server error")
         }
     }
@@ -207,13 +178,13 @@ class RequestHandler(
         }
     }
 
-    private suspend fun storeUploadedFile(body: String, prefix: String, contentType: String): Path {
+    private fun storeUploadedFile(body: String, prefix: String, contentType: String): Path {
         val tempFile = of(oscalDir.toString(), "$prefix-${System.nanoTime()}${getExtensionFromContentType(contentType)}")
         Files.writeString(tempFile, body)
         return tempFile
     }
 
-    suspend fun handleValidateFileUpload(ctx: RoutingContext) {
+    fun handleValidateFileUpload(ctx: RoutingContext) {
         launch {
             try {
                 logger.info("Handling validate content request")
@@ -242,7 +213,7 @@ class RequestHandler(
         }
     }
 
-    suspend fun handleQueryFileUpload(ctx: RoutingContext) {
+    fun handleQueryFileUpload(ctx: RoutingContext) {
         launch {
             try {
                 logger.info("Handling query content request")
@@ -271,7 +242,7 @@ class RequestHandler(
         }
     }
 
-    suspend fun handleResolveFileUpload(ctx: RoutingContext) {
+    fun handleResolveFileUpload(ctx: RoutingContext) {
         launch {
             try {
                 logger.info("Handling resolve content request")
@@ -300,7 +271,7 @@ class RequestHandler(
         }
     }
 
-    suspend fun handleConvertFileUpload(ctx: RoutingContext) {
+    fun handleConvertFileUpload(ctx: RoutingContext) {
         launch {
             try {
                 logger.info("Handling convert content request")
@@ -328,5 +299,4 @@ class RequestHandler(
             }
         }
     }
-
 }
